defore deploy by master

+++ pages/admin/users.js
@@ -85,6 +85,9 @@ function AdminUsers() {
   const { enqueueSnackbar } = useSnackbar();
 
   const deleteHandler = async (userId) => {
+    if (!window.confirm('Are you sure?')) {
+      return;
+    }
     try {
       dispatch({ type: 'DELETE_REQUEST' });
       await axios.delete(`/api/admin/users/${userId}`, {
